Move DebugInfo::populateSourceMap to DebugInfoNonLean.cpp for lean build
Summary:
The lean build does not include hermesSourceMap, but DebugInfo.cpp
referenced SourceMapGenerator::addFunctionOffsets via populateSourceMap.
This causes link failures on MSVC which requires all symbols to be
resolved even for unreferenced code.

Move populateSourceMap to a new DebugInfoNonLean.cpp file that is only
compiled into the full hermesHBCBackend library, not the lean version.

Also move FunctionDebugInfoDeserializer from anonymous namespace to
hermes::hbc::detail namespace in the header so it can be shared between
DebugInfo.cpp and DebugInfoNonLean.cpp.

Reviewed By: lavenzg

Differential Revision: D89912026

fbshipit-source-id: 2d43d7564f17cafb45b1c5a68608dc6ec3052e6d
