+++ lib/BCGen/HBC/DebugInfo.cpp
@@ -8,109 +8,10 @@
 #include "hermes/BCGen/HBC/DebugInfo.h"
 
 #include "hermes/BCGen/HBC/ConsecutiveStringStorage.h"
-#include "hermes/SourceMap/SourceMapGenerator.h"
 
-using namespace hermes;
-using namespace hbc;
+namespace hermes::hbc {
 
-namespace {
-
-/// A type used to iteratively deserialize function debug info.
-struct FunctionDebugInfoDeserializer {
-  /// Construct a deserializer that begins deserializing at \p offset in \p
-  /// data. It will deserialize until the function's debug info is finished
-  /// (address delta = -1) at which point isDone() will return true. The offset
-  /// of the next section can be obtained via getOffset().
-  /// isDone() will not be true until after the first call to next().
-  FunctionDebugInfoDeserializer(llvh::ArrayRef<uint8_t> data, uint32_t offset)
-      : data_(data), offset_(offset) {
-    functionIndex_ = decode1Int();
-    current_.line = decode1Int();
-    current_.column = decode1Int();
-    current_.envIdx = decode1Int();
-  }
-
-  /// \return the next debug location, or None if we reach the end.
-  /// Sample usage: while (auto loc = fdid.next()) {...}
-  OptValue<DebugSourceLocation> next() {
-    assert(!done_ && "next() called after done");
-
-    auto addressDelta = decode1Int();
-    if (addressDelta == -1) {
-      done_ = true;
-      return llvh::None;
-    }
-    current_.address += addressDelta;
-
-    // ldelta encoding: bits 3..32 contain the line delta.
-    // Bit 0 indicates the presence of location information.
-    // Bit 1 indicates the presence of statement information.
-    // Bit 2 indicates the presence of envIdx information.
-
-    int64_t lineDelta = decode1Int();
-    if ((lineDelta & 1) == 0) {
-      // No location information here, but not done yet.
-      return llvh::None;
-    }
-
-    int64_t columnDelta = decode1Int();
-    int64_t statementDelta = 0;
-    if ((lineDelta & 2) != 0) {
-      statementDelta = decode1Int();
-    }
-
-    int64_t envIdxDelta = 0;
-    if ((lineDelta & 4) != 0) {
-      envIdxDelta = decode1Int();
-    }
-
-    lineDelta >>= 3;
-
-    current_.line += lineDelta;
-    current_.column += columnDelta;
-    current_.statement += statementDelta;
-    current_.envIdx += envIdxDelta;
-    return current_;
-  }
-
-  /// \return whether we're done reading the debug info for this function.
-  /// next() must not be called when isDone() is true.
-  bool isDone() const {
-    return done_;
-  }
-
-  /// \return the offset of this deserializer in the data.
-  uint32_t getOffset() const {
-    return offset_;
-  }
-
-  /// \return the index of the function being deserialized.
-  uint32_t getFunctionIndex() const {
-    return functionIndex_;
-  }
-
-  /// \return the current source location.
-  const DebugSourceLocation &getCurrent() const {
-    return current_;
-  }
-
- private:
-  /// LEB-decode the next int.
-  int64_t decode1Int() {
-    int64_t result;
-    offset_ += readSignedLEB128(data_, offset_, &result);
-    return result;
-  }
-
-  llvh::ArrayRef<uint8_t> data_;
-  uint32_t offset_;
-
-  uint32_t functionIndex_;
-  DebugSourceLocation current_;
-
-  bool done_ = false;
-};
-} // namespace
+using FunctionDebugInfoDeserializer = detail::FunctionDebugInfoDeserializer;
 
 OptValue<uint32_t> DebugInfo::getFilenameForAddress(
     uint32_t debugOffset) const {
@@ -282,52 +183,6 @@ void DebugInfo::disassembleFilesAndOffsets(llvh::raw_ostream &OS) const {
      << "  end of debug source table\n\n";
 }
 
-void DebugInfo::populateSourceMap(
-    SourceMapGenerator *sourceMap,
-    std::vector<uint32_t> &&functionOffsets,
-    uint32_t segmentID) const {
-  // Since our bytecode is not JavaScript, we interpret the source map in a
-  // creative way: each bytecode module is represented as a line, and bytecode
-  // addresses in the file are represented as column offsets.
-  // Our debug information has a function start and then offsets within the
-  // function, but the source map will do its own delta encoding, so we provide
-  // absolute addresses to the source map.
-  auto segmentFor = [&](const DebugSourceLocation &loc,
-                        uint32_t offsetInFile,
-                        uint32_t debugOffset) {
-    SourceMap::Segment segment;
-    segment.generatedColumn = loc.address + offsetInFile;
-    assert(loc.line >= 1 && "line numbers in debug info must be 1-based");
-    assert(loc.column >= 1 && "column numbers in debug info must be 1-based");
-    segment.representedLocation = SourceMap::Segment::SourceLocation(
-        sourceMap->getSourceIndex(
-            getUTF8FilenameByID(*getFilenameForAddress(debugOffset))),
-        loc.line - 1,
-        loc.column - 1);
-    return segment;
-  };
-
-  std::vector<SourceMap::Segment> segments;
-  llvh::ArrayRef<uint8_t> locsData = sourceLocationsData();
-  uint32_t offset = 0;
-  while (offset < locsData.size()) {
-    FunctionDebugInfoDeserializer fdid(locsData, offset);
-    uint32_t offsetInFile = functionOffsets[fdid.getFunctionIndex()];
-    segments.push_back(segmentFor(fdid.getCurrent(), offsetInFile, offset));
-    while (!fdid.isDone()) {
-      if (auto loc = fdid.next()) {
-        segments.push_back(segmentFor(*loc, offsetInFile, offset));
-      }
-      // Don't add a segment for the debug info that has no location,
-      // because we won't be reporting location info on instructions that have
-      // no location information for the source map to translate.
-    }
-    offset = fdid.getOffset();
-  }
-  sourceMap->addMappingsLine(std::move(segments), segmentID);
-  sourceMap->addFunctionOffsets(std::move(functionOffsets), segmentID);
-}
-
 uint32_t DebugInfoGenerator::addScopingInfo(
     const DebugScopingInfo &scopingInfo) {
   auto &scopingInfoTable = debugInfo_.getScopingInfoTable();
@@ -430,3 +285,5 @@ void DebugInfoGenerator::generate() && {
 
   debugInfo_.getFilenameTableMut().appendFilenamesToStorage();
 }
+
+} // namespace hermes::hbc
